+++ server/db/update_scripts/38_update_v7.5.0_v8.0.0.sql
@@ -810,6 +810,95 @@ CREATE TABLE IF NOT EXISTS `apiRequestLogs` (
   `error_message` TEXT DEFAULT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
 
+DELIMITER //
+DROP PROCEDURE IF EXISTS add_foreign_key //
+CREATE PROCEDURE add_foreign_key(param_table VARCHAR(100), fk_name VARCHAR(100), fk_column VARCHAR(100), fk_references VARCHAR(200))
+BEGIN	
+    SET @sqlstmt = (SELECT IF(
+		(
+			SELECT COUNT(*)
+            FROM information_schema.TABLE_CONSTRAINTS 
+			WHERE `table_schema` = DATABASE()
+			AND `table_name` = param_table
+            AND `constraint_name` = fk_name
+		) > 0,
+        "SELECT 'The foreign key already exists in the table'",
+        CONCAT('ALTER TABLE ', param_table, ' ADD CONSTRAINT ', fk_name, ' FOREIGN KEY (', fk_column, ') REFERENCES ', fk_references, ' ON DELETE CASCADE;')
+    ));
+	PREPARE st FROM @sqlstmt;
+	EXECUTE st;
+	DEALLOCATE PREPARE st;	
+END
+
+//
+
+DELIMITER ;
+
+DELIMITER //
+DROP PROCEDURE IF EXISTS add_index //
+CREATE PROCEDURE add_index(
+    param_table VARCHAR(100), 
+    param_index_name VARCHAR(100), 
+    param_index_column VARCHAR(1000),
+    param_is_unique BOOLEAN
+)
+BEGIN	
+    SET @sqlstmt = (SELECT IF(
+		(
+			SELECT COUNT(*)
+            FROM information_schema.STATISTICS 
+			WHERE `table_schema` = DATABASE()
+			AND `table_name` = param_table
+            AND `index_name` = param_index_name
+		) > 0,
+        "SELECT 'The index already exists in the table'",
+        CONCAT(
+            'CREATE ', 
+            IF(param_is_unique, 'UNIQUE ', ''),
+            'INDEX ', 
+            param_index_name, 
+            ' ON `', 
+            param_table, 
+            '` (`', 
+            param_index_column, 
+            '`);'
+        )
+    ));
+	PREPARE st FROM @sqlstmt;
+	EXECUTE st;
+	DEALLOCATE PREPARE st;	
+END
+
+//
+
+DELIMITER ;
+
+
+DELIMITER //
+DROP PROCEDURE IF EXISTS drop_index //
+CREATE PROCEDURE drop_index(param_table VARCHAR(100), param_index_name VARCHAR(100))
+BEGIN	
+    SET @sqlstmt = (SELECT IF(
+		(
+			SELECT COUNT(*)
+            FROM information_schema.STATISTICS 
+			WHERE `table_schema` = DATABASE()
+			AND `table_name` = param_table
+            AND `index_name` = param_index_name
+		) > 0,        
+        CONCAT('ALTER TABLE `', param_table, '` DROP INDEX ', param_index_name),
+        "SELECT 'The index does not exists in the table'"
+    ));
+	PREPARE st FROM @sqlstmt;
+	EXECUTE st;
+	DEALLOCATE PREPARE st;	
+END
+
+//
+
+DELIMITER ;
+
+
 
 
 -- shoudl remove is_fluid from container style