+++ CHANGELOG.md
@@ -14,6 +14,7 @@
  - add `slug` type in the `router`
  - migrate tables: `actions`,`activityType`,`styleType` and `userStatus` to `lookups`
  - add `ApiRequestLog` table where all the requests to the `cms-api` are logged
+ - foreign keys only cascade on delete and not on update
 
 ### Bugfix
  - fix the way that the data is decoded in the `formUserInput` controller